Add modules required for metaport to work, move interfaces to core package, add ABIs
