+++ packages/core/package.json
@@ -6,7 +6,7 @@
   "types": "dist/index.d.ts",
   "type": "module",
   "scripts": {
-    "build": "tsc && bun build ./src/index.ts --outdir ./dist --target node",
+    "build": "tsc && bun build ./src/index.ts --outdir ./dist --target browser",
     "type-check": "tsc --noEmit"
   },
   "author": "SKALE Labs",
