+++ packages/core/src/types/index.ts
@@ -30,7 +30,8 @@ export {
     CategoriesMap,
     AppWithChainAndName
 } from './ChainsMetadata'
-export * as staking from './staking'
+export * as st from './staking'
+export * as mp from './metaport'
 
 export type AddressType = `0x${string}`
 export type Size = 'xs' | 'sm' | 'md' | 'lg'
