+++ packages/metaport/package.json
@@ -49,6 +49,7 @@
     "@typescript-eslint/eslint-plugin": "5.60.0",
     "@vitejs/plugin-react": "4.0.4",
     "@vitest/coverage-v8": "0.34.1",
+    "ajv": "^8.17.1",
     "autoprefixer": "10.4.14",
     "eslint": "8.46.0",
     "eslint-config-prettier": "9.0.0",
@@ -75,6 +76,7 @@
     "vitest": "0.34.1"
   },
   "dependencies": {
+    "@/core": "file:../core",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@mui/icons-material": "^5.15.6",
