+++ packages/metaport/src/components/Metaport/Metaport.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import Metaport from './Metaport'
 
-import { METAPORT_CONFIG } from '../../metadata/metaportConfigTestnet'
+import { METAPORT_CONFIG } from '../../meta/testnetConfig'
 METAPORT_CONFIG.mainnetEndpoint = import.meta.env.VITE_MAINNET_ENDPOINT
 METAPORT_CONFIG.projectId = import.meta.env.VITE_WC_PROJECT_ID
