+++ packages/metaport/src/components/TransactionData/TransactionData.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables';
+@use '../../styles/variables' as *;
 
 .transactionDataIcon {
   width: 30px;
