+++ packages/metaport/src/metadata/addresses/testnet.json
@@ -1,0 +1,10 @@
+{
+  "message_proxy_mainnet_address": "0x682ef859e1cE314ceD13A6FA32cE77AaeCE98e28",
+  "linker_address": "0x840A1BCa22c6D67Ed1076A1018b404EbcaA8a10c",
+  "community_pool_address": "0xF263049E4D7b331154077e30dD4de72F779E9554",
+  "deposit_box_eth_address": "0xe1A65b7677866A80B6e4263469FAC22ddc26F4b8",
+  "deposit_box_erc20_address": "0x4121c9218a65e12f027714AE956Aea7B9C220dAE",
+  "deposit_box_erc721_address": "0x024cdFF8Cd5cF334243B2fC2fd3eA1Fc1C9ebec5",
+  "deposit_box_erc1155_address": "0xb990E0D1F115398D7E6A8E4269e68b9034366280",
+  "deposit_box_erc721_with_metadata_address": "0x81caAfd5DF89d9DA44054c0a0a305E35DD7e958E"
+}
