+++ packages/metaport/src/styles/cmn.module.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@use './variables' as *;
 
 .flex {
     display: flex;
