+++ packages/metaport/src/styles/styles.module.scss
@@ -1,4 +1,4 @@
-@import './variables';
+@use './variables' as *;
 
 .paper {
     padding: 10px;
